Refactored Class and Working on Theme Selector
