+++ app/src/main/AndroidManifest.xml
@@ -5,8 +5,8 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
     <uses-permission android:name="android.permission.STORAGE" />
-    <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="com.android.vending.BILLING" />
 
     <!-- permission required to use Alarm Manager -->
@@ -16,72 +16,72 @@
 
     <supports-screens
         android:anyDensity="true"
-        android:smallScreens="true"
-        android:normalScreens="true"
         android:largeScreens="true"
-        android:xlargeScreens="true"
-        android:resizeable="true" />
+        android:normalScreens="true"
+        android:resizeable="true"
+        android:smallScreens="true"
+        android:xlargeScreens="true" />
 
     <application
         android:allowBackup="true"
+        android:fullBackupContent="false"
         android:icon="@mipmap/ic_say_it"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/BlueYellowStyle.Theme"
-        android:fullBackupContent="false">
-
-        <meta-data android:name="com.google.android.gms.version" android:value="@integer/google_play_services_version" />
-        <meta-data android:name="android.max_aspect" android:value="2.1" />
-        <!--Include the AdActivity configChanges and theme. -->
-        <activity android:name="com.google.android.gms.ads.AdActivity"
+        android:theme="@style/BlueYellowStyle.Theme">
+        <meta-data
+            android:name="com.google.android.gms.version"
+            android:value="@integer/google_play_services_version" />
+        <meta-data
+            android:name="android.max_aspect"
+            android:value="2.1" />
+        <!-- Include the AdActivity configChanges and theme. -->
+        <activity
+            android:name="com.google.android.gms.ads.AdActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenLayout|uiMode|screenSize|smallestScreenSize"
             android:theme="@android:style/Theme.Translucent" />
-
-
         <activity
-            android:name="com.cesarsk.say_it.ui.MainActivity"
+            android:name=".ui.activities.MainActivity"
             android:launchMode="singleTask"
             android:screenOrientation="portrait">
-
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-
             <intent-filter android:label="@string/app_name">
                 <action android:name="android.intent.action.VIEW" />
+
                 <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.BROWSABLE" />
                 <!-- Accepts URIs that begin with "http://www.example.com/gizmos” -->
-                <data android:scheme="http"
+                <data
                     android:host="www.lucacesaranoblog.wordpress.com"
-                    android:pathPrefix="/" />
+                    android:pathPrefix="/"
+                    android:scheme="http" />
             </intent-filter>
         </activity>
-
         <activity
-            android:name="com.cesarsk.say_it.ui.SettingsActivity"
+            android:name=".ui.activities.SettingsActivity"
             android:label="@string/title_activity_settings"
-            android:parentActivityName="com.cesarsk.say_it.ui.MainActivity"
+            android:parentActivityName=".ui.activities.MainActivity"
             android:screenOrientation="portrait"
             android:theme="@style/BlueYellowStyle.PreferenceTheme" />
         <activity
-            android:name="com.cesarsk.say_it.ui.PlayActivity"
-            android:parentActivityName="com.cesarsk.say_it.ui.MainActivity"
+            android:name=".ui.activities.PlayActivity"
+            android:parentActivityName=".ui.activities.MainActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name="com.cesarsk.say_it.ui.SearchActivity"
-            android:parentActivityName="com.cesarsk.say_it.ui.MainActivity"
+            android:name=".ui.activities.SearchActivity"
+            android:parentActivityName=".ui.activities.MainActivity"
             android:screenOrientation="portrait" />
-        <activity android:name="com.cesarsk.say_it.ui.FileTextActivity" />
+        <activity android:name=".ui.activities.FileTextActivity" />
 
         <receiver
-            android:name="com.cesarsk.say_it.NotificationReceiver"
+            android:name=".notifications.NotificationReceiver"
             android:enabled="true" />
-
         <receiver
-            android:name="com.cesarsk.say_it.NotificationBootReceiver"
+            android:name=".notifications.NotificationBootReceiver"
             android:enabled="false">
             <intent-filter>
                 <action android:name="android.intent.action.BOOT_COMPLETED" />
