+++ app/src/main/java/com/cesarsk/say_it/ui/activities/FileTextActivity.java
@@ -1,4 +1,4 @@
-package com.cesarsk.say_it.ui;
+package com.cesarsk.say_it.ui.activities;
 
 import android.app.Activity;
 import android.content.Intent;
