+++ app/src/main/java/com/cesarsk/say_it/ui/components/FavoritesAdapter.java
@@ -16,8 +16,8 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.cesarsk.say_it.R;
-import com.cesarsk.say_it.ui.MainActivity;
-import com.cesarsk.say_it.ui.PlayActivity;
+import com.cesarsk.say_it.ui.activities.MainActivity;
+import com.cesarsk.say_it.ui.activities.PlayActivity;
 import com.cesarsk.say_it.ui.fragments.FavoritesFragment;
 import com.cesarsk.say_it.utility.SayItPair;
 import com.cesarsk.say_it.utility.UtilitySharedPrefs;
