+++ app/src/main/java/com/cesarsk/say_it/ui/fragments/RecordingsFragment.java
@@ -1,11 +1,7 @@
 package com.cesarsk.say_it.ui.fragments;
 
 
-import android.app.ActivityOptions;
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
-import android.content.Intent;
 import android.graphics.Canvas;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
@@ -23,13 +19,10 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.cesarsk.say_it.ui.MainActivity;
+import com.cesarsk.say_it.ui.activities.MainActivity;
 import com.cesarsk.say_it.R;
 import com.cesarsk.say_it.ui.components.RecordingsAdapter;
-import com.cesarsk.say_it.ui.PlayActivity;
-import com.cesarsk.say_it.utility.SayItPair;
 import com.cesarsk.say_it.utility.Utility;
-import com.cesarsk.say_it.utility.UtilityDictionary;
 import com.cesarsk.say_it.utility.UtilityRecordings;
 
 import java.io.File;
@@ -37,8 +30,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
-
-import uk.co.deanwild.materialshowcaseview.MaterialShowcaseView;
 
 
 /**
